ci(build): publish rolling minor and pinned patch versions (#22)
Starting with BeeGFS 8, package repositories exist for each major.minor version, and patch releases
are released as needed into these repos. This commit updates the release workflow so a major.minor
tag is pushed which determines the major.minor version of the images, but the patch version is
determined by looking at the package versions in the package repo.

This also removes the beegfs-all container image to avoid mismatched tagging.